@@ -16,7 +16,8 @@
     "gulp-sass": "^2.0.4",
     "gulp-util": "^3.0.6",
     "imports-loader": "^0.6.4",
-    "react": "^0.13.3",
+    "react": "^15.4.0",
+    "react-dom": "^15.4.0",
     "webpack": "^1.12.0",
     "webpack-stream": "^2.1.0"
   },
